+++ tests/test_06_math/test_618_clipping.py
@@ -4,100 +4,10 @@
 import pytest
 from ezdxf.math import Vec2
 from ezdxf.math.clipping import (
-    clip_polygon_2d,
     ClippingPolygon2d,
     ClippingRect2d,
 )
 from ezdxf.render.forms import circle
-
-
-@pytest.fixture
-def rect():
-    return [(-1, -1), (1, -1), (1, 1), (-1, 1)]
-
-
-@pytest.fixture
-def overlapping():  # overlapping
-    return [(0, 0), (2, 0), (2, 2), (0, 2)]
-
-
-@pytest.fixture
-def inside():  # complete inside
-    return [(0, 0), (0.5, 0), (0.5, 0.5), (0, 0.5)]
-
-
-@pytest.fixture
-def outside():  # complete outside
-    return [(1, 1), (2, 1), (2, 2), (1, 2)]
-
-
-def test_subject_do_overlap_clipping_rect(rect, overlapping):
-    result = clip_polygon_2d(rect, overlapping)
-    assert len(result) == 4
-    assert Vec2(0, 0) in result
-    assert Vec2(1, 0) in result
-    assert Vec2(1, 1) in result
-    assert Vec2(0, 1) in result
-
-
-def test_subject_is_inside_rect(rect, inside):
-    result = clip_polygon_2d(rect, inside)
-    assert len(result) == 4
-    for v in inside:
-        assert any(r.isclose(v) for r in result) is True
-
-
-def test_clockwise_oriented_clipping_rect(rect, inside):
-    rect.reverse()
-    result = clip_polygon_2d(rect, inside)
-    assert len(result) == 4
-    for v in inside:
-        assert any(r.isclose(v) for r in result) is True
-
-
-def test_subject_is_outside_rect(rect, outside):
-    result = clip_polygon_2d(rect, outside)
-    assert len(result) == 0
-
-
-def test_circle_outside_rect(rect):
-    c = circle(16, 3)
-    result = clip_polygon_2d(rect, c, ccw_check=False)
-    assert len(result) == 4
-    for v in rect:
-        assert any(r.isclose(v) for r in result) is True
-
-
-def test_circle_inside_rect(rect):
-    c = Vec2.list(circle(16, 0.7))
-    result = clip_polygon_2d(rect, c, ccw_check=False)
-    assert len(result) == 16
-    for v in c:
-        assert any(r.isclose(v) for r in result) is True
-
-
-def test_rect_outside_circle(rect):
-    c = circle(16, 0.7)
-    result = clip_polygon_2d(c, rect, ccw_check=False)
-    assert len(result) == 16
-    for v in c:
-        assert any(r.isclose(v) for r in result) is True
-
-
-def test_rect_inside_circle(rect):
-    c = circle(16, 3)
-    result = clip_polygon_2d(c, rect, ccw_check=False)
-    assert len(result) == 4
-    for v in rect:
-        assert any(r.isclose(v) for r in result) is True
-
-
-def test_clip_a_single_line(rect):
-    line = [(0.5, 0), (2, 0)]
-    result = clip_polygon_2d(rect, line, ccw_check=False)
-    assert len(result) == 2
-    assert result[0].isclose((1.0, 0))
-    assert result[1].isclose((0.5, 0))
 
 
 class TestClippingSingleLines:
@@ -180,6 +90,89 @@ class TestClippingPolyline:
         assert p0[1].isclose((3, 2))
         assert p1[0].isclose((5, 2))
         assert p1[1].isclose((7, 0))
+
+
+class TestClippingPolygons:
+    @pytest.fixture(scope="class", params=["polygon", "rect"])
+    def clipper(self, request):
+        if request.param == "polygon":
+            return ClippingPolygon2d(
+                Vec2.list([(-1, -1), (1, -1), (1, 1), (-1, 1)])
+            )
+        else:
+            return ClippingRect2d(Vec2(-1, -1), Vec2(1, 1))
+
+    @pytest.fixture
+    def rect(self):
+        return Vec2.list([(-1, -1), (1, -1), (1, 1), (-1, 1)])
+
+    @pytest.fixture
+    def overlapping(self):  # overlapping
+        return Vec2.list([(0, 0), (2, 0), (2, 2), (0, 2)])
+
+    @pytest.fixture
+    def inside(self):  # complete inside
+        return Vec2.list([(0, 0), (0.5, 0), (0.5, 0.5), (0, 0.5)])
+
+    @pytest.fixture
+    def outside(self):  # complete outside
+        return Vec2.list([(1, 1), (2, 1), (2, 2), (1, 2)])
+
+    def test_subject_do_overlap_clipping_rect(self, clipper, overlapping):
+        result = clipper.clip_polygon(overlapping)
+        assert len(result) == 4
+        assert Vec2(0, 0) in result
+        assert Vec2(1, 0) in result
+        assert Vec2(1, 1) in result
+        assert Vec2(0, 1) in result
+
+    def test_subject_is_inside_rect(self, clipper, inside):
+        result = clipper.clip_polygon(inside)
+        assert len(result) == 4
+        for v in inside:
+            assert any(r.isclose(v) for r in result) is True
+
+    def test_clockwise_oriented_clipping_rect(self, rect, inside):
+        rect.reverse()
+        clipper = ClippingPolygon2d(rect)
+        result = clipper.clip_polygon(inside)
+        assert len(result) == 4
+        for v in inside:
+            assert any(r.isclose(v) for r in result) is True
+
+    def test_subject_is_outside_rect(self, clipper, outside):
+        result = clipper.clip_polygon(outside)
+        assert len(result) == 0
+
+    def test_circle_outside_rect(self, clipper, rect):
+        c = Vec2.generate(circle(16, 3))
+        result = clipper.clip_polygon(c)
+        assert len(result) == 4
+        for v in rect:
+            assert any(r.isclose(v) for r in result) is True
+
+    def test_circle_inside_rect(self, clipper):
+        c = Vec2.list(circle(16, 0.7))
+        result = clipper.clip_polygon(c)
+        assert len(result) == 16
+        for v in c:
+            assert any(r.isclose(v) for r in result) is True
+
+    def test_rect_outside_circle(self, rect):
+        c = Vec2.list(circle(16, 0.7))
+        clipper = ClippingPolygon2d(c)
+        result = clipper.clip_polygon(rect)
+        assert len(result) == 16
+        for v in c:
+            assert any(r.isclose(v) for r in result) is True
+
+    def test_rect_inside_circle(self, rect):
+        c = Vec2.list(circle(16, 3))
+        clipper = ClippingPolygon2d(c)
+        result = clipper.clip_polygon(rect)
+        assert len(result) == 4
+        for v in rect:
+            assert any(r.isclose(v) for r in result) is True
 
 
 if __name__ == "__main__":