+++ NEWS.md
@@ -7,7 +7,9 @@ Version 0.18.1b0 - dev
 
 - Release notes: https://ezdxf.mozman.at/release-v0-18.html
 - NEW: `ezdxf.render.forms.turtle()`, create a 2D polyline vertices by 
-  turtle-graphic like commands  
+  turtle-graphic like commands
+- BUGFIX: CPython implementation of `Vec2()` was not immutable at inplace 
+  operations `+=`, `-=`, `*=` and `/=` like the Cython implementation
 
 Version 0.18 - 2022-07-29
 -------------------------