+++ NEWS.md
@@ -7,6 +7,8 @@ Version 0.18.1b0 - dev
 
 - Release notes: https://ezdxf.mozman.at/release-v0-18.html
 - NEW: improved hatch pattern support for the `drawing` add-on
+- NEW: `drawing` add-on got basic `VIEWPORT` rendering (only top-views), 
+  supported by the `PyQtBackend` and the `PillowBackend`  
 - NEW: `ezdxf.render.forms.turtle()` function to create 2D polyline vertices by 
   turtle-graphic like commands
 - NEW: sub-command `ezdxf pillow` to draw and convert DXF files by `Pillow`